requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-pubsub
pandas~=1.3.5
appstoreconnect~=0.9.1